o  ID related updates to Continuum facade, Configuration service and notifier Wagon
git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/branches/id-refactor@497920 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-api/src/main/java/org/apache/maven/continuum/configuration/ConfigurationService.java
@@ -55,10 +55,10 @@ public interface ConfigurationService
 
     void setDeploymentRepositoryDirectory( File deploymentRepositoryDirectory );
 
-    String getBuildOutput( int buildId, int projectId )
+    String getBuildOutput( long buildId, long projectId )
         throws ConfigurationException;
 
-    File getBuildOutputFile( int buildId, int projectId )
+    File getBuildOutputFile( long buildId, long projectId )
         throws ConfigurationException;
 
     // ----------------------------------------------------------------------
